disparando alarm para o led não travar o thread principal

+++ build.sh
@@ -1,7 +1,10 @@
 #!/bin/sh -e
 
-rm -rf ./build
-mkdir ./build
+if [ "$1" = "clean" ]; then
+  rm -rf ./build
+  mkdir ./build
+fi
+
 cd ./build
 cmake -DPICO_SDK_PATH=../pico-sdk -DPICO_BOARD=pico ..
 make hello_world
